==> notebooks/Fig5/panel_A/2.control_enrichment/1.QC_filter_and_impute.ipynb <==
# Preprocessing

Pre-processing includes QC filtering, Log2 intensity transformation, removing rows that have insufficient data, and global imputation  
In this notebook, we processes the IP columns, and the N/O/C columns will be separately processed in another notebook

import os, sys
from pathlib import Path
import pandas as pd

script_path = Path.cwd().parent.parent.parent.parent / "script"
data_path = Path.cwd().parent.parent.parent.parent / "data"
sys.path.append(str(script_path))

from pyseus import basic_processing as ip

### Define input and output files

%store -r fig5_timestamp FIG5_USE_FROZEN
if FIG5_USE_FROZEN:
    raise Exception("USE_FROZEN is true, you probably want to skip enrichment and proceed from 3.aligned_umap")
timestamp = fig5_timestamp

maxQuant_csv = "proteinGroups.txt" # name of the MaxQuant output file
# see data/MaxQuant/readme.md for more information about this file

outprefix = f"{timestamp}_QC_filter_impute"

outdir = Path.cwd() / "output"

preprocessing_out = outdir / f"preprocessing"
os.makedirs(preprocessing_out, exist_ok=True)

preprocessing_log = preprocessing_out / f"logs"
os.makedirs(preprocessing_log, exist_ok=True)

### Read the infected enrichment table
The purpose of this is to get a list of infected samples, so we can subset the uninfected samples to create matching sets of samples

filename = f'{timestamp}_inf_enrichment_table_NOC_prop.csv'
infected_enrichment_path = Path.cwd().parent / "1.infected_enrichment" / "output" / "enrichment_and_volcano_tables" / filename 

try:
    infected_enrichments = pd.read_csv(infected_enrichment_path, header=[0, 1], index_col=0)
except FileNotFoundError:
    print(f"File {infected_enrichment_path} not found.\n please run the infected_enrichment notebooks first.")
except pd.errors.ParserError:
    print(f"There was an error parsing the CSV file at {infected_enrichment_path}.")
except Exception as e:
    print(f"An unexpected error occurred: {e}")

infected_samples = infected_enrichments["sample"].columns
# predict the names of uninfected samples from infected samples
uninfected_pred = [
    i.split("_")[0] for i in infected_samples if not i.startswith("NOC")
] + ["Uninfected-Cytosol", "Uninfected-Nuclear", "Uninfected-Organelle", "Uninfected-Proteome"]

print(f"the number of infected samples is {len(infected_samples)}")

### Load proteingroups table

# import the protein groups table
pgroups_path = data_path / "MaxQuant" / maxQuant_csv

try:
    pgroups = pd.read_csv(pgroups_path, index_col=None, low_memory=False, sep='\t')
except FileNotFoundError:
    print(f"File {pgroups_path} not found.\nPlease note that this file is 670MB and is not included in the repository.")
except pd.errors.ParserError:
    print(f"There was an error parsing the CSV file at {pgroups_path}.")
except Exception as e:
    print(f"An unexpected error occurred: {e}")

# define exclusion columns to remove non-relavent samples from the mass spectrometry master file
exclusion_cols1 = [x for x in list(pgroups.columns) if 'Infected' in x or "_infected" in x or '-infected' in x] # exclude infected samples
exclusion_cols2 = [x for x in list(pgroups.columns) if 'Harsh' in x or "Unsorted" in x] # exclude harsh and unsorted samples
exclusion_cols3 = [x for x in list(pgroups.columns) if 'brefeldin' in x or 'arsenite' in x]  # exclude brefeldin and arsenite samples

# select uninfected samples that matches the infected samples
exclusion_cols4 = [x for x in list(pgroups.columns) if not any([i in x.replace("_", "-") for i in uninfected_pred])] 

exclusion_cols = sorted(list(set(exclusion_cols1 + exclusion_cols2 + exclusion_cols3 + exclusion_cols4)))

# fix capitalization errors in column names
exclusion_cols = [i for i in exclusion_cols if i.startswith("LFQ")]

# write exclusion_col to file
with open(preprocessing_log / f"{outprefix}_exclusion_cols.txt", 'w') as f:
    for item in exclusion_cols:
        if item.startswith("LFQ"):
            f.write(f"{item}\n")

# use the inverse of the exclusion columns to select the inclusion columns
inclusion_cols = [x for x in list(pgroups.columns) if x not in exclusion_cols]
# write inclusion column names to file for examination
count = 0
with open(preprocessing_log / f"{outprefix}_inclusion_cols.txt", 'w') as f:
    for item in inclusion_cols:
        if item.startswith("LFQ"):
            f.write("%s\n" % item)

print(f"the number of columns in the protein groups table is {int(count/3)}, not including the NOC columns")

# sample naming rule: a single underscore is used to separate the sample name from the replicate number, for example SampleA_1, SampleA_2 and SampleA_3 are replicates of SampleA
# the sample name is used to group replicates together
# experimental conditions (eg infection) are part of sample names, and therefore gets a dash instead of underscore

re = ["infected", '_InfectedOC43', 'WT_HEK', '_Unsorted', '_HarshLysis', '_rep1', '_rep2', '_rep3', r'( \d{2})_'] # regular expression to search for
replacement_re=["Infected",'-infected', 'WT', '-unsorted', '-HarshLysis', '_1', '_2', '_3', r'\1-'] # replacement regular expression

# the method will reformat strings by the order given in the previous list
# this is a dry run and intended to test whether the renaming is working as intended
sample_cols = ip.sample_rename(inclusion_cols, RE=re, replacement_RE=replacement_re, repl_search=False) 

# subset pggroups
pgroups = pgroups[inclusion_cols]

# rename samples
meta_cols = ["Protein IDs", "Majority protein IDs", "Gene names"]
preprocessing = ip.RawTables(
    proteingroup=pgroups,
    sample_cols=list(pgroups),
    file_designated=True,
    info_cols=meta_cols,
    intensity_type="LFQ intensity",  # Duo: type selection can't be use together with sample_cols selection
)

# rename files
preprocessing.rename_columns(RE=re, replacement_RE=replacement_re, repl_search=False) 

# you can access renamed table and renamed cols by the following class objects
_ = preprocessing.renamed_table
_ = preprocessing.sample_cols

# write sample column names to file for examination
with open(os.path.join(preprocessing_log / f"{outprefix}__sample_cols.txt"), "w") as f:
    for item in preprocessing.sample_cols:
        if item.startswith("LFQ"):
            f.write("%s\n" % item)

### Pre-processing steps
Pre-processing includes QC filtering, Log2 intensity transformation, removing rows that have insufficient data,
and global imputation

# filter table based on MaxQuant qc
preprocessing.filter_table(select_intensity=True, verbose=True)

# transform intensities
preprocessing.transform_intensities()

# group replicates and remove insufficient rows
preprocessing.group_replicates(reg_exp=r'(.*)_\d+$')


# require proteins to have at leaset two valid values in the proteome samples
preprocessing.remove_invalid_rows_custom(["UnInfected_Proteome"])

preprocessing.remove_invalid_rows()

# save preimpute table
preprocessing.preimpute_table.to_csv(preprocessing_out / f"{outprefix}_preimpute_table.csv")

# impute for NaN values. Here we are using global imputation 
preprocessing.bait_impute(distance=1.8, width=0.3, local=False)

Save the imputed table

# the pre-processing is done, you can save/continue using the below class object
_ = preprocessing.bait_imputed_table

# print out the column names to file for examination
preprocessing.bait_imputed_table.columns.to_frame().to_csv(preprocessing_out / f"{outprefix}_imputed_table_cols.csv")

# write preprocessed pg table to a csv file (this is required for next steps)
preprocessing.bait_imputed_table.to_csv(preprocessing_out / f"{outprefix}_imputed_table.csv")

# write preprossed pg table to tabular file (for Perseus) 
#preprocessing.bait_imputed_table.to_csv(preprocessing_out / f"{outprefix}_imputed_table.tab", sep='\t')

# write the filtered table to file (for NOC processing)
preprocessing.filtered_table.to_csv(preprocessing_out / f"{outprefix}_filtered_table.csv")

# view the bait_imputed_table
preprocessing.bait_imputed_table
